updated time type

+++ messages.go
@@ -116,7 +116,7 @@ type MotionMessage struct {
 	Header  MessageHeader `json:"header,omitempty"`
 	Device  string        `json:"device,omitempty"`
 	Channel string        `json:"channel,omitempty"`
-	TimeMs  uint          `json:"time_ms,omitempty"`
+	TimeMs  int64         `json:"time_ms,omitempty"`
 }
 
 type TemperatureMessage struct {
